+++ src/Inventory/Types/LocationDetails.php
@@ -8,17 +8,31 @@
  * is updated. You've been warned!
  */
 
-namespace DTS\eBaySDK\Fulfillment\Types;
+namespace DTS\eBaySDK\Inventory\Types;
 
 /**
  *
+ * @property \DTS\eBaySDK\Inventory\Types\Address $address
+ * @property \DTS\eBaySDK\Inventory\Types\GeoCoordinates $geoCoordinates
  */
-class OrderFulfillmentStatus extends \DTS\eBaySDK\Types\BaseType
+class LocationDetails extends \DTS\eBaySDK\Types\BaseType
 {
     /**
      * @var array Properties belonging to objects of this class.
      */
     private static $propertyTypes = [
+        'address' => [
+            'type' => 'DTS\eBaySDK\Inventory\Types\Address',
+            'repeatable' => false,
+            'attribute' => false,
+            'elementName' => 'address'
+        ],
+        'geoCoordinates' => [
+            'type' => 'DTS\eBaySDK\Inventory\Types\GeoCoordinates',
+            'repeatable' => false,
+            'attribute' => false,
+            'elementName' => 'geoCoordinates'
+        ]
     ];
 
     /**