+++ src/Trading/Types/GroupValidationRulesType.php
@@ -8,17 +8,24 @@
  * is updated. You've been warned!
  */
 
-namespace DTS\eBaySDK\Fulfillment\Types;
+namespace DTS\eBaySDK\Trading\Types;
 
 /**
  *
+ * @property integer $MinRequired
  */
-class FulfillmentInstructionsType extends \DTS\eBaySDK\Types\BaseType
+class GroupValidationRulesType extends \DTS\eBaySDK\Types\BaseType
 {
     /**
      * @var array Properties belonging to objects of this class.
      */
     private static $propertyTypes = [
+        'MinRequired' => [
+            'type' => 'integer',
+            'repeatable' => false,
+            'attribute' => false,
+            'elementName' => 'MinRequired'
+        ]
     ];
 
     /**
@@ -32,6 +39,10 @@ class FulfillmentInstructionsType extends \DTS\eBaySDK\Types\BaseType
 
         if (!array_key_exists(__CLASS__, self::$properties)) {
             self::$properties[__CLASS__] = array_merge(self::$properties[get_parent_class()], self::$propertyTypes);
+        }
+
+        if (!array_key_exists(__CLASS__, self::$xmlNamespaces)) {
+            self::$xmlNamespaces[__CLASS__] = 'xmlns="urn:ebay:apis:eBLBaseComponents"';
         }
 
         $this->setValues(__CLASS__, $childValues);