+++ src/Trading/Types/RecommendationsType.php
@@ -13,6 +13,7 @@ namespace DTS\eBaySDK\Trading\Types;
 /**
  *
  * @property string $CategoryID
+ * @property \DTS\eBaySDK\Trading\Types\ProductIdentifiersType[] $ProductIdentifiers
  * @property \DTS\eBaySDK\Trading\Types\NameRecommendationType[] $NameRecommendation
  * @property boolean $Updated
  */
@@ -27,6 +28,12 @@ class RecommendationsType extends \DTS\eBaySDK\Types\BaseType
             'repeatable' => false,
             'attribute' => false,
             'elementName' => 'CategoryID'
+        ],
+        'ProductIdentifiers' => [
+            'type' => 'DTS\eBaySDK\Trading\Types\ProductIdentifiersType',
+            'repeatable' => true,
+            'attribute' => false,
+            'elementName' => 'ProductIdentifiers'
         ],
         'NameRecommendation' => [
             'type' => 'DTS\eBaySDK\Trading\Types\NameRecommendationType',