+++ test/Trading/Types/ProductIdentifiersTypeTest.php
@@ -8,22 +8,22 @@
  * is updated. You've been warned!
  */
 
-namespace DTS\eBaySDK\Test\Order\Types;
+namespace DTS\eBaySDK\Test\Trading\Types;
 
-use DTS\eBaySDK\Order\Types\ShippingAddress_0;
+use DTS\eBaySDK\Trading\Types\ProductIdentifiersType;
 
-class ShippingAddress_0Test extends \PHPUnit_Framework_TestCase
+class ProductIdentifiersTypeTest extends \PHPUnit_Framework_TestCase
 {
     private $obj;
 
     protected function setUp()
     {
-        $this->obj = new ShippingAddress_0();
+        $this->obj = new ProductIdentifiersType();
     }
 
     public function testCanBeCreated()
     {
-        $this->assertInstanceOf('\DTS\eBaySDK\Order\Types\ShippingAddress_0', $this->obj);
+        $this->assertInstanceOf('\DTS\eBaySDK\Trading\Types\ProductIdentifiersType', $this->obj);
     }
 
     public function testExtendsBaseType()